Fail on CMake run if SSL enabled but OpenSSL not found
A CMake run now fails in case OpenSSL is not found when the local
PostgreSQL installation has been compiled with SSL enabled.

The reason why it is best to fail the CMake run in this case is
because USE_OPENSSL will be defined and 1 in `pg_config.h` and thus we
will compile with SSL support.