Add URI type and parser
Telemetry is now supported by a URI type and parser. This enables easy
testing against various endpoints (localhost or remote), using either
HTTP or HTTPS, as well as non-standard ports.

The telemetry endpoint GUC now is a full URI, including resource path.
This means the path need no longer be hard coded.

This also fixes a number of test issues. For instance, in the
telemetry test the escape char `E` was passed on as part of the
response string when set as a variable with `\set`. This was not
detected before because the response parser didn't parse the start of
the response properly.