Make launcher track all DB state its DB hash table
This change ensures we track all DBs in the launcher's hash table and
not only the ones that have the extension installed and have a running
scheduler. This obviates the need for extra state tracking in
alternative data structures for, e.g., capturing create DB events from
template databases that already have the extension installed.