Respect refresh window when bucketing invalidations
When refreshing a continuous aggregate in a given window, it will only
refresh the buckets that are invalidated in that window. When
processing invalidations, they are always expanded to align with and
include full buckets. However, the expansion must also respect the
original refresh window, which can either align with buckets or chunk
boundaries.

Fixes #2609