Fix crash when user mapping has no user name
When a user mapping is created without a user name in the options a
crash will occur when querying a distributed hypertable.

This change fixes the crash by using the "current user" when no user
option exists (i.e., no user mapping is done). This makes it possible
to create user mappings that only provide the password option.