Avoid SPI connect in nested procedure
When a procedure internally executes multiple transactions (such as in
the experimental `copy_chunk` and `move_chunk`), one needs to check
for active SPI connections in order to maintain the SPI state across
transactions. This happens when a multi-transaction procedure is
called within another procedure; the outer procedure is executed using
an SPI context that needs to remain once the inner multi-transactional
procedure completes.

An issue related to maintaining the SPI context was previously fixed
by opening another SPI context. But this inner context is never used.

This change removes the unnecessary SPI context.