Make SQLsmith workflow use workflow_dispatch instead of build matrix
Refactor the SQLsmith workflow to take its input from the
workflow_dispatch event instead of relying on a build matrix.

Using workflow_dispatch lets the user run a particular configuration,
e.g., by specificying a particular seed to generate queries that make
it easier to reproduce a failure or crash.

Also, upload query logs, which makes it possible to find the exact
query that caused a failure.