Add function to alter data nodes
Add a new function, `alter_data_node()`, which can be used to change a
data node's configuration on the access node, including connection
information like host, port, and database.

In addition, the function allows marking a node as "unavailable",
which means other data nodes will be used for queries. The function
finds all the chunks for which the unavailable data node is the
"default" query target and sets another data node as the default. For
this to work, all the chunks on the unavailable data node must have a
replica on another data node. If some chunks are not replicated, a
warning will be raised.

When a data node is available again, the function should allow
switching back to once again use the data node for queries. However,
this part is not yet implemented and is left for a future change.

Closes #2104