Support calendar-based chunking
Hypertables can now be configured to create chunks that align with the
start and end of days, weeks, months, and years. This includes days and
months that vary in length due to daylight savings or month of the year.

The alignment happens according to the current time zone setting of the
database. If the time zone is changed, existing hypertables will adapt
to the new setting although existing chunk partitions remain as is
(i.e., aligned with the old time zone).