Fix ALTER TABLE spurious printout
If `create_compress_chunk()` is called as part of a utility statement,
a spurious printout will happen because
`ts_alter_table_with_event_trigger()` uses a `List` node rather than a
proper statement which causes `EventTriggerAlterTableEnd()` to not
recognize the node and print out an error message.

This is fixed by calling `AlterTableInternal` instead of
`ts_alter_table_with_event_trigger()` for utility statements and make sure that
`EventTriggerAlterTableStart()` and `EventTriggerAlterTableEnd()` are
called for non-utility statements calling `create_compress_chunk()`
with a dummy statement.

Fixes #997