Fix speculative insert when there is no conflict
Add tests for speculative inserts when there is no conflict. During
such inserts, decompression currently happens in order to populate the
unique index that conflicts are checked against.

With the compression TAM, this decompression is no longer necessary
because a unique index covers both compressed and non-compressed
data. However, the decompression is not disabled as part of this
change in order to not affect the current implementation of
compression. This unnecessary compression will be removed later.

To handle speculative inserts, it is necessary to implement the
index_delete_tuples() callback in the TAM. This callback is used by
the indexAM to ask the tableAM about potential tuples to delete. The
tuple TIDs are passed to the callback as an array, and may include
TIDs referencing both compressed and non-compressed tuples. Therfore,
the compression TAM's version of index_delete_tuples() needs to split
The TID array into two: one for the compressed relation and one for
the non-compressed one, and then call the corresponding heapAM
callback on each relation individually.