Use index paths for compressed chunks
When compression table access method is used, we should use index paths
if transparent decompression is disabled. Otherwise, index scans will
not be used and the chunk will always use a sequence scan.

This commit will keep the existing indexes if transparent decompression
is disabled allowing, for example, plans containing indexes to be
generated when the compressed table access method is used.