Support recompression in compression TAM
A chunk using compression TAM can now be recompressed using
`recompress_chunk`. One issue with supporting recompression is to
ensure it is possible to scan only "non-compressed" data. This is done
via a scan key flag, telling the TAM that it should only return
non-compressed data.

It is also necessary to set the internal chunk state to "partially
compressed". This doesn't work properly in legacy compression if data
is inserted directly into the chunk. This has been fixed for
compression TAM by setting the chunk status to partially compressed at
the end of a transaction that did inserts.