Fix UUID tests
Fixing the sanitizer tests that flags up unaligned access that
I added in the unit tests to test exactly that.

Fixing the flakyness of the UUID filtering test that filtered
timestamps on a table that has v4 (random) data, so the timestamp
returned may or may not match the timesptamp filter.

Disable-check: force-changelog-file