Add option to zero-initialize generated v7 UUIDs
To be able to filter v7 UUIDs based on time, add an option to
zero-initialize UUIDs created from timestamps. The function can be
used in queries to, e.g., find all UUIDs with a timestamp greater or
equal to a given value like so:

SELECT uuid, temp
WHERE uuid >= uuid_v7_from_timestamptz('2025-01-01', zeroed => true);

The function creates a UUIDv7 with all bits except the timestamp set
to zero, thus representing the "first" UUID with the given timestamp
in time order.

Also tweak the volatility of UUID-related functions to be more
restrictive so that the query planner can optimize them more easily.