Add function to generate UUIDv7 from unixtime
The new function is created for internal use by breaking up the
function to generate a UUIDv7 from a timestamp with time zone in two
pieces: first convert the timestamp with time zone to unixtime, then
call the unixtime function in second step.

Creating a UUIDv7 from unixtime is useful when implementing native
partitioning support for UUIDv7, since the internal time used for
constraint metadata is in unixtime.