Fix update race with BGW jobs
There are re-occurring crashes in background worker (BGW) tests caused
by an assertion that checks for duplicate jobs with the same primary key
ID. Since it is a primary key, such duplicate jobs should not exist.
However, the BGW code scans its metadata tables with the default Scanner
SnapshotSelf snapshot. Under this snapshot, it is possible to read both
current tuples and new versions of the tuples that are in progress of
being updated by another process but are not yet committed.

The code to handle job metadata should really use a MVCC-compatible
snapshot in order to ensure proper visibility. However, switching
snapshot used for all functions is a big undertaking since some
functionality (like updates and deletes of metadata) rely on being
able to see tuples in a non-MVCC way.

Therefore, this fix changes some read functions to use an active
snapshot, while functions that mutate remains as is.

In summary, the following changes are made:

- Use the "active" transaction snapshot to read job metadata
- Add a test that triggers the duplicate row assertion when
  an MVCC-compatible snapshots is _not_ used.